+++ lib/onlyoffice_digitalocean_wrapper/ssh_checker/ssh_checker_exceptions.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module OnlyofficeDigitaloceanWrapper
+  # Class raised if droplet ssh is not up for defined timeout
+  class SshCheckerSshUpTimeout < StandardError
+  end
+end