+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### New features
 
-* `DigitalOceanWrapper#ssh_up?` and `DigitalOceanWrapper#wait_until_ssh_up` methods
+* `SshChecker#ssh_up?` and `SshChecker#wait_until_ssh_up` methods
 
 ## 0.7.0 (2021-07-29)